Rename signal to interrupt
