+++ src/main.rs
@@ -4,6 +4,7 @@
 
 mod console;
 mod exit_code;
+mod interrupt;
 mod lab;
 mod log_file;
 mod mutate;
@@ -11,7 +12,6 @@ mod options;
 mod outcome;
 mod output;
 mod run;
-mod signal;
 mod source;
 mod textedit;
 mod visit;
@@ -27,8 +27,8 @@ use argh::FromArgs;
 use path_slash::PathExt;
 
 // Imports of public names from this crate.
+use crate::interrupt::was_interrupted;
 use crate::options::Options;
-use crate::signal::was_interrupted;
 use crate::source::SourceTree;
 
 /// Find inadequately-tested code that can be removed without any tests failing.
@@ -83,7 +83,7 @@ fn main() -> Result<()> {
         .show_all_logs(args.all_logs)
         .show_times(!args.no_times);
     let options = Options::from(&args);
-    signal::install_handler();
+    interrupt::install_handler();
     if args.list {
         let mutations = source_tree.mutations()?;
         if args.json {
